// projects/share-files-dialog/src/lib/dialog.enum.ts
export enum PermissionTypes {
  owner = 'owner',
  reader = 'reader',
  writer = 'writer',
  multiple = 'multiple',
}

export enum SortingPermissionTypes {
  'owner' = 0,
  'reader' = 1,
  'writer' = 2,
  'multiple' = 3,
}

export enum FileTypes {
  DOCUMENT = 'document',
  VIDEO = 'video',
  AUDIO = 'audio',
  IMAGE = 'image',
}